CoK19, Formations, Allies, and MORE validation!
When choosing Forces, the [Main Army] includes points for the intended Roster (game) size. For 0-1495 (zero? RIP!) point Rosters, there is a cap of 1 max duplicate Hero/Mon/WarEng. For 1500+ point Rosters, the cap changes depending on the exact amount of points currently spent, and is 100% properly validating things like 1995 and 2995 point games that don't want to hit the next breakpoint due to balance issues, or other reasons. When including {Allied Units} in the 1500+ point bracket, the minimum points required to reach 1500 for your Main Army are lowered by 5 points for every 5 points of anything purchased in the Allied Force, until you reach the 25% cap for 1500 total points (375 of Allies, 1125 of Main), where the minimum disappears. You are never required to set an actual "points limit" in your Rosters. With this KoW data it is purely for convenience. Nothing relies on it.